+++ packages/subgraph/src/handler.ts
@@ -10,30 +10,37 @@ import {
   SubtreeCommit,
   Nullifier,
 } from "../generated/schema";
-import { getIdWithEntityIndex, getTotalLogIndex } from "./utils";
+import {
+  getTotalLogIndex,
+  getTotalEntityIndex,
+  toPadded32BArray,
+} from "./utils";
 
 export function handleJoinSplit(event: JoinSplitProcessed): void {
   const totalLogIndex = getTotalLogIndex(event);
   const joinSplit = event.params.joinSplit;
 
   // first old note's nullifier
-  let id = getIdWithEntityIndex(totalLogIndex, 0);
-  const nullifierA = new Nullifier(id);
+  let idx = getTotalEntityIndex(totalLogIndex, 0);
+  const nullifierA = new Nullifier(toPadded32BArray(idx));
+  nullifierA.idx = idx;
   nullifierA.nullifier = event.params.oldNoteANullifier;
   nullifierA.save();
 
   // second old note's nullfier
-  id = getIdWithEntityIndex(totalLogIndex, 1);
-  const nullifierB = new Nullifier(id);
+  idx = getTotalEntityIndex(totalLogIndex, 1);
+  const nullifierB = new Nullifier(toPadded32BArray(idx));
+  nullifierB.idx = idx;
   nullifierB.nullifier = event.params.oldNoteBNullifier;
   nullifierB.save();
 
   // unpack first new note
-  id = getIdWithEntityIndex(totalLogIndex, 2);
+  idx = getTotalEntityIndex(totalLogIndex, 2);
+  let id = toPadded32BArray(idx);
 
   const encryptedNoteA = new EncryptedNote(id);
-
   const newNoteAEncrypted = joinSplit.newNoteAEncrypted;
+  encryptedNoteA.idx = idx;
   encryptedNoteA.ownerH1X = newNoteAEncrypted.owner.h1X;
   encryptedNoteA.ownerH1Y = newNoteAEncrypted.owner.h1Y;
   encryptedNoteA.ownerH2X = newNoteAEncrypted.owner.h2X;
@@ -50,16 +57,19 @@ export function handleJoinSplit(event: JoinSplitProcessed): void {
   encryptedNoteA.save();
 
   const newNoteA = new EncodedOrEncryptedNote(id);
+  newNoteA.idx = idx;
   newNoteA.merkleIndex = event.params.newNoteAIndex;
   newNoteA.encryptedNote = id;
   newNoteA.save();
 
   // unpack second new note
-  id = getIdWithEntityIndex(totalLogIndex, 3);
 
+  idx = getTotalEntityIndex(totalLogIndex, 2);
+  id = toPadded32BArray(idx);
   const encryptedNoteB = new EncryptedNote(id);
 
   const newNoteBEncrypted = joinSplit.newNoteBEncrypted;
+  encryptedNoteB.idx = idx;
   encryptedNoteB.ownerH1X = newNoteBEncrypted.owner.h1X;
   encryptedNoteB.ownerH1Y = newNoteBEncrypted.owner.h1Y;
   encryptedNoteB.ownerH2X = newNoteBEncrypted.owner.h2X;
@@ -76,6 +86,7 @@ export function handleJoinSplit(event: JoinSplitProcessed): void {
   encryptedNoteB.save();
 
   const newNoteB = new EncodedOrEncryptedNote(id);
+  newNoteB.idx = idx;
   newNoteB.merkleIndex = event.params.newNoteBIndex;
   newNoteB.encryptedNote = id;
   newNoteB.save();
@@ -84,12 +95,14 @@ export function handleJoinSplit(event: JoinSplitProcessed): void {
 export function handleRefund(event: RefundProcessed): void {
   const totalLogIndex = getTotalLogIndex(event);
 
-  const id = getIdWithEntityIndex(totalLogIndex, 0);
+  const idx = getTotalEntityIndex(totalLogIndex, 0);
+  const id = toPadded32BArray(idx);
   const newNote = new EncodedOrEncryptedNote(id);
   const encodedNote = new EncodedNote(id);
 
   const refundAddr = event.params.refundAddr;
 
+  encodedNote.idx = idx;
   encodedNote.ownerH1X = refundAddr.h1X;
   encodedNote.ownerH1Y = refundAddr.h1Y;
   encodedNote.ownerH2X = refundAddr.h2X;
@@ -102,16 +115,19 @@ export function handleRefund(event: RefundProcessed): void {
 
   newNote.merkleIndex = event.params.merkleIndex;
   newNote.note = id;
+  newNote.idx = idx;
   newNote.save();
 }
 
 export function handleSubtreeUpdate(event: SubtreeUpdate): void {
   const totalLogIndex = getTotalLogIndex(event);
 
-  const id = getIdWithEntityIndex(totalLogIndex, 0);
+  const idx = getTotalEntityIndex(totalLogIndex, 0);
+  const id = toPadded32BArray(idx);
   const commit = new SubtreeCommit(id);
 
   commit.newRoot = event.params.newRoot;
   commit.subtreeIndex = event.params.subtreeIndex;
+  commit.idx = idx;
   commit.save();
 }